Removed collate_tpms.R dependency on vignette_config.yaml
Replaced:

* Use of `--yaml` command-line option.
* Parsing of YAML to get `dir_out` and sample names from `fq_files`
* `# ORFs <- readDNAStringSet(yps$orf_fasta) %>% names # if ORF list from fasta`

with

* `--dir_out` command-line option.
* `--orf_fasta` command-line option (optional).
* Positional arguments with sample names.

This both removes the dependency on a YAML file allows different files to be collated without the need to edit such a YAML file.

A caller now invokes, for example:

```
$ Rscript --vanilla rscripts/collate_tpms.R --dir_out vignette/output/ WTnone WT3AT NotHere
```

or

```
$ Rscript --vanilla rscripts/collate_tpms.R --dir_out vignette/output/ --orf_fasta vignette/input/yeast_YAL_CDS_w_250utrs.fa WTnone WT3AT NotHere
```

prep_riboviz.py updated to use these new parameters. run-vignette.md updated to show sample invocation.

Relates to #33.