simdata renamed from example
Simulated fastq data for umis/barcodes/adaptors is now called "simdata" and not "example".
That is because it is not a real example, but actually a simulation.
We may want to create "example" config files indicating how to deal with new data, rather than tied to the simulated data.
Also, build_indices=FALSE for simdata, because rebuilding indices for a tiny test is a waste of time