Resolved issue of non-use of cached results with `-resume`
The previous implementation of `trim5pMismatches` was as follows:

```
process trim5pMismatches {
    tag "${sample_id}"
    publishDir "${params.dir_tmp}/${sample_id}", \
        mode: 'copy', overwrite: true
    errorStrategy 'ignore'
    input:
        env PYTHONPATH from workflow.projectDir
        tuple val(sample_id), file(sam) from orf_map_sams
    output:
	tuple val(sample_id), file("orf_map_clean.sam") \
            into clean_orf_map_sams
        tuple val(sample_id), file("trim_5p_mismatch.tsv") \
            into trim_summary_tsvs
    shell:
        """
        python -m riboviz.tools.trim_5p_mismatch -m 2 \
            -i ${sam} -o orf_map_clean.sam -s trim_5p_mismatch.tsv
        """
}
```

Attempts to resume the workflow and reuse cached results, via the `-resume` flag, always caused this process to be reexecuted - the cached results were not reused. For example:

```
[b4/cb053f] Cached process > hisat2ORF (WTnone)
[2a/b8edf8] Cached process > hisat2ORF (WT3AT)
[65/f9061c] Submitted process > trim5pMismatches (WTnone)
[c4/a99951] Submitted process > trim5pMismatches (WT3AT)
```

Using the `nextflow run` `-dump-hashes` flag showed that the value of `workflow.projectDir` (for me, `/home/ubuntu/riboviz`) which is assigned to `PYTHONPATH`, has a different hash value each run. For example:

```
  06e49a1561ac1c4cc6b56c3f0fd092ab [java.lang.String] PYTHONPATH
  79f1cea846383a3712a9a1ef445edb0b [sun.nio.fs.UnixPath] /home/ubuntu/riboviz
```
```
  06e49a1561ac1c4cc6b56c3f0fd092ab [java.lang.String] PYTHONPATH
  fc0b39720d66636e6fa63169b4a858ca [sun.nio.fs.UnixPath] /home/ubuntu/riboviz
```

From "How is the hash calculated on input files?" in [Demystifying Nextflow resume](https://www.nextflow.io/blog/2019/demystifying-nextflow-resume.html)
 the hash of a file is derived from the complete file path, the file size, the last modified timestamp. As Nextflow is run within `/home/ubuntu/riboviz`, its timestamp would change per run and so the hash of `workflow.projectDir` would change. It is this that triggers the reinvocation of the process. See also [Troubleshooting Nextflow resume](https://www.nextflow.io/blog/2019/troubleshooting-nextflow-resume.html).

Removing the `env ...` definition and, instead, defining `PYTHONPATH` in the shell command allowed for cached results to be reused (since there is no hashing of `workflow.projectDir`):

```
    shell:
        """
        PYTHONPATH=${workflow.projectDir} python -m riboviz.tools.trim_5p_mismatch -m 2 \
            -i ${sam} -o orf_map_clean.sam -s trim_5p_mismatch.tsv
        """
```

Alternatively, defining `PYTHONPATH` in the call to `nextflow run` also allowed for cached results to be reused (likewise):

```console
$ PYTHONPATH=/home/ubuntu/riboviz nextflow run prep_riboviz.nf -params-file vignette/vignette_config.yaml -ansi-log false  -resume
```

Alternatively, converting `workflow.projectDir` to a string also allows for cached results to be reused:

```
        env PYTHONPATH from workflow.projectDir.toString()
```

In this case it is the string value of `workflow.projectDir` that is hashed and this always evaluates to the same value. For example:

```
  8289d52801c5c6fece0763679a15274d [java.lang.String] /home/ubuntu/riboviz
```

This fix was adopted, as closest to the original code, and applied to all Python-invoking processes. It was tested by running each of:

```console
$ nextflow run prep_riboviz.nf -params-file vignette/vignette_config.yaml -ansi-log false
$ nextflow run prep_riboviz.nf -params-file vignette/simdata_umi_config.yaml -ansi-log false
$ nextflow run prep_riboviz.nf -params-file vignette/simdata_multiplex_config.yaml -ansi-log false
```

then rerunning with `-resume` and ensuring the output showed that all cached results were reused.

Longer term having the Python scripts installed within the current Python environment would remove the need to worry about the `PYTHONPATH`.