Extended support for parameter and input file validation
Added default values for:

* buffer
* count_threshold
* dataset
* dir_index
* dir_output
* dir_tmp
* do_pos_sp_nt_freq
* is_riboviz_gff
* max_read_length
* min_read_length
* primary_id
* rpf
* stop_in_cds

Added existence checks for parameters and files:

* orf_fasta_file
* orf_gff_file
* rrna_fasta_file
* fq_files
* multiplex_fq_files
* sample_sheet
* asite_disp_length_file
* codon_positions_file
* features_file
* t_rna_file

Added existence checks for parameters:

* adapters
* dir_in
* orf_index_prefix
* rrna_index_prefix

For any failed check for a parameter, `exit 1, 'Message'` is used, not `error` so that a Nextflow-specific error with a Nextflow file line number is avoided.

Fixed bug in `if (multiplex_file.exists() || true)`, removed `|| true`.