Completed `riboviz.test.test_sample_sheet`
This includes a test `test_get_non_zero_deplexed_samples` that is designed to fail on the current, flawed, implementation of `get_non_zero_deplexed_samples` (arising from the use of bitwise 'and' in Pandas). Once a fix is committed to `get_non_zero_deplexed_samples` (to use parentheses to ensure the 'and' is evaluated correctly) then the test should pass.

Relates to #214.