Support optional parameters in staticHTML process
Updated `staticHTML` to build a string, in Groovy, to ensure that
only those parameters for which values are provided are added
to the R fragment.

If the conditionals are false then the parameters are
not provided. This is consistent with `generateStatsFigs`.

Fixed bug caused by introduction of asite_disp_length_txt1 and
asite_disp_length_txt2 by removing their use. These were redundant.

Updated `generateStatsFigs` to:

* No longer declare optional outputs for its optional `.tsv` files.
* `touch` the files for every optional output after invocation of
  `generate_stats_figs.R`. This ensures that `generateStatsFigs`
  always produces outputs which will ensure that `staticHTML` will run.
  While empty versions of the optional output files are created, these
  will not be passed to
  `rmarkdown::render('${workflow.projectDir}/rmarkdown/AnalysisOutputs.Rmd'...)`
  as the conditionals added earlier will prevent that.