Added prep_riboviz.nf support for RIBOVIZ_SAMPLES, RIBOVIZ_ORGANISMS, RIBOVIZ_DATA environment variables
`prep_riboviz.nf`:

* Gets values of RIBOVIZ_SAMPLES, RIBOVIZ_ORGANISMS, RIBOVIZ_DATA environment
  variables and assigns to corresponding Nextflow variables.
* For any undefined environment variable, the value "." is used.
* If the path pointed at by an environment variable does not exist then
  the workflow exits with an error.
* Defines a mapping from tokens '${RIBOVIZ_SAMPLES}',
  '${RIBOVIZ_ORGANISMS}, '${RIBOVIZ_DATA}' to the Nextflow variables
  above.

A `replace_tokens` function that replaces a token in string with
a corresponding substring can be used to replace the above tokens
with the values of the corresponding Nextfloe values.

Invocation of `replace_tokens` when getting RiboViz file parameters
has still to be done.

A fragment of commented out code, to be removed, has examples of
how this function can be used.