Removed redundant number of data columns check from test_bam_to_h5.R
Removed:
```
  num_data_cols <- h5_stop_codon_pos[3] + buffer
  expect_equal(ncol(h5_data), num_data_cols,
    info = paste0(sequence,
      ": Number of data columns != stop_codon_pos[3] + buffer"))
```
since this is covered by following checks:
```
  expect_equal(ncol(h5_data), utr3_end,
    info = paste0(sequence,
      ": Number of data columns != GFF UTR3 final nt position"))
  expect_equal(ncol(h5_data), bam_hdr_sequence_seq_length,
    info = paste0(sequence,
      ": Number of data columns != BAM sequence length"))
```
Also removed since it `buffer` should not be used if validating
outputs if `is_riboviz_gff` was `true`.

`utr3_end` was already defined earlier in the test to hold the
expected value regardless of whether `is_riboviz_gff` was `true`
or `false`.