Updated test_bam_to_h5.R to extract primary_id and secondary_id
The test now correctly uses primary_id to access information from GFF
and H5 file. Formerly it used the sequence name. This is now
consistent with bam_to_h5.R.

secondary_id is extracted and used to validate names in H5 but
nothing more as yet.

This revealed an issue with the tests which is logged in a TODO. The
code currently assumes that the sequence names and primary IDs in
the GFF file are the same for each feature (as that was the case
in all the GFF files consulted when writing the test).

In fact, the H5 file has entries named after the feature names,
but we need the sequence names to access information about the
feature from the BAM file. So, from the GFF above we need a map
from feature name to sequence name, for example, which we can
both get the unique sequence names and for cross-checking against
the BAM file then iterating through to validate each
sequence/feature pair in turn.