Deleted rscripts/tests/testthat/test_generate_stats_figs.R
It is very prototypical, fails to run, and only tests the number of
arguments given and for existence of an input file.
Deleting it for now.
When tests come to be written for `generate_stats_figs.R` it
can be resurrected.
`rscripts/tests/testthat/test_bam_to_h5.R` has a working example
of `testthat` tests for an R script.