Removed .here file and added troubleshooting
Undid commit e8b98c92a77e43d69e9a6ee3754062a6c5625167 which added `.here` file.

While this fixed the AnalysisOutputs.md bug in using 'here' when no
`.git` folder was present it was found to cause the same problem
when run on EDDIE. It is unclear as to why. As most users are likely
to use a clone of RiboViz the `.git` folder will be present.

For those using a .zip from GitHub or Figshare, troubleshooting
has been added to docs/user/prep-riboviz-run-nextflow.md for such
users to add `.here` themselves and, if that is unsuccessful to
use a clone from GitHub and to add their experiences to issue #352.