Updated prep_riboviz.nf staticHTML to use intermediate directory
Updated call to rmarkdown::render in staticHTML to set directory
for intermediate files to be $PWD i.e. work directory for task.
This means that sample-specific intermediate files are created
in sample-specific directories rather than in rmarkdown/ (leading
to intermediate files for different samples colliding as they
all are prefixed by AnalysisOutputs but have no sample-specific
identifier in their name).